Clean up code by pulling it out of setup
